\name{RawToChar-class}
\Rdversion{1.1}
\docType{class}
\alias{RawToChar}
\alias{RawToChar-class}

\title{Class "RawToChar"}

\description{

  A \code{\linkS4class{Consumer}}-class to convert \code{raw} (binary)
  records to \code{char}, applying \code{rawToChar} to each record.

}

\usage{RawToChar(...)}

\arguments{
  \item{...}{Arguments passed to the \code{\linkS4class{Consumer}}-class
    constructors.}
}

\section{Constructors}{
  Use \code{RawToChar} to construct instances of this class.
}

\section{Fields}{There are no fields unique to this class.}

\section{Methods}{There are no methods unique to this class.}

\author{Martin Morgan \url{mtmorgan@fhcrc.org}}

\seealso{\code{\link{stream}}}

\examples{showClass("RawToChar")}

\keyword{classes}
