\name{Streamer-class}
\Rdversion{1.1}
\docType{class}
\alias{Streamer-class}

\title{Class defining methods shared by all classes}
\description{

  A virtual base class from which all classes in the Streamer package
  derive.

}

\section{Methods}{

  See \code{Stream}, \code{Producer}, and \code{Consumer} Methods.

}

\section{Internal Class Fields and Methods}{

  Internal fields of this class are are described with, e.g.,
  \code{getRefClass("Streamer")$fields}.

  Internal methods of this class are described with
  \code{getRefClass("Streamer")$methods()} and
  \code{getRefClass("Streamer")$help()}.

}

\author{Martin Morgan \url{mtmorgan@fhcrc.org}}

\seealso{

  \code{\link{Streamer-package}}, \code{\linkS4class{Consumer}-class},
  \code{\linkS4class{Producer}-class}, \code{\linkS4class{Stream}-class}.

}

\examples{showClass("Streamer")}

\keyword{classes}
